Split measure tool UI into sub components
